Fixed propagation of PETSC solver options for positive definite systems in water flow.
The flag 'is_positive_definite' does not propagate from a LinSys to its Schur complement, thus it has to be manually set to the complement.
For parallel run some test did not converge with CG+ICC (one should modify the amount of fill), so default options for parallel are kept at BCGS+ILU.

Fix #289.